organized the csv files and merged them together
put all the csv files into a folder called "Labeled csv" and I also created a notebook called "labeled_csv_merging" to merge all the csv files together. This merged file is also uploaded as "ccombined_labeled_data.csv" which we can use as a starting point for all of the machine learning that follows.